Add files via upload
Here are the souce files for the assignment. Uploading the whole project means way too many files and apparently I cannot do that.
Generate Report is not implemented yet. UI not really completed, missing some display options. Pay Utilities is modeled as a transfer between a client's account and the utilitie's account. The id's for the utilities will be displayed in the UI.

I created tests for all methods in services and repositories and all seems to be working fine. I would appreciate if you could take a look and tell me if there is anything I should consider modifying.